fix: 修复消费明细API avg_latency_ms类型转换错误
将ConsumptionSummary.AvgLatencyMs从int改为float64，
解决PostgreSQL AVG()返回浮点数导致的scan error

+++ model/log_stats.go
@@ -6,10 +6,10 @@ import (
 
 // ConsumptionSummary 消费汇总
 type ConsumptionSummary struct {
-	TotalQuota    int64 `json:"total_quota"`
-	TotalRequests int64 `json:"total_requests"`
-	TotalTokens   int64 `json:"total_tokens"`
-	AvgLatencyMs  int   `json:"avg_latency_ms"`
+	TotalQuota    int64   `json:"total_quota"`
+	TotalRequests int64   `json:"total_requests"`
+	TotalTokens   int64   `json:"total_tokens"`
+	AvgLatencyMs  float64 `json:"avg_latency_ms"`
 }
 
 // ModelConsumption 按模型分组的消费
